Revert "small tweaks to scanning"
This reverts commit 820de1363debfde5583ec7070342a56abe3702f5.

This doesn't work: we still need to scan target and mouseover in case we learn the GUID but don't change the target/mouseover.

The vb stuff was just unnecessary, the time between learning the GUID and setting the icon is realistically at worst 5 seconds long, more like 3. Further, if you can get that synced that implies someone else is running DBM and found it. They will have already set the icon and you no longer need to.

The target/mouseover mix-up is already fixed in another commit